+++ apis/vara-eth/src/provider/index.ts
@@ -1,2 +1,3 @@
 export * from './http.js';
+export * from './validator-pool.js';
 export * from './ws.js';