+++ apis/vara-eth/src/types/provider.ts
@@ -1,3 +1,5 @@
+import type { Address } from 'viem';
+
 export type ISubscriptionCallback<Error = unknown, Result = unknown> = (
   error: Error,
   result: Result,
@@ -14,3 +16,14 @@ export interface IVaraEthProvider {
     callback: ISubscriptionCallback<Error, Result>,
   ): Promise<() => void>;
 }
+
+export interface IVaraEthValidatorPoolProvider extends IVaraEthProvider {
+  readonly isPool: true;
+  readonly validatorAddresses: Address[];
+  setActiveValidator(address: Address): void;
+  removeValidator(address: Address): Promise<void>;
+  addValidator(address: Address, url: string): Promise<void>;
+}
+
+export type WsUrl = `ws://${string}` | `wss://${string}`;
+export type HttpUrl = `http://${string}` | `https://${string}`;